+++ Exiled.Events/EventArgs/StayingOnEnvironmentalHazardEventArgs.cs
@@ -21,12 +21,10 @@ namespace Exiled.Events.EventArgs
         /// </summary>
         /// <param name="player"><inheritdoc cref="Player"/></param>
         /// <param name="environmentalHazard"><inheritdoc cref="EnvironmentalHazard"/></param>
-        /// <param name="isAllowed"><inheritdoc cref="IsAllowed"/></param>
-        public StayingOnEnvironmentalHazardEventArgs(Player player, EnvironmentalHazard environmentalHazard, bool isAllowed = true)
+        public StayingOnEnvironmentalHazardEventArgs(Player player, EnvironmentalHazard environmentalHazard)
         {
             Player = player;
             EnvironmentalHazard = environmentalHazard;
-            IsAllowed = isAllowed;
         }
 
         /// <summary>
@@ -38,10 +36,5 @@ namespace Exiled.Events.EventArgs
         /// Gets the environmental hazard that the player is staying on.
         /// </summary>
         public EnvironmentalHazard EnvironmentalHazard { get; }
-
-        /// <summary>
-        /// Gets or sets a value indicating whether or not the player should be affected by the environmental hazard.
-        /// </summary>
-        public bool IsAllowed { get; set; }
     }
 }
